Driver
Most of Driver methods implemented
(not tested)

+++ TaxiService/TaxiService/Models/Base/CommentBase.cs
@@ -24,6 +24,6 @@ namespace TaxiService.Models
         //public virtual RideBase CommentRide { get; set; }
 
         [Range(0,5)]
-        public ushort Stars { get; set; }
+        public int Stars { get; set; }
     }
 }
